refractor(aicode-utils): combine-packages merge scaffold-generator to aicode-utils
