+++ packages/aicode-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@agiflowai/aicode-utils",
-  "description": "Utility functions and helpers for AI-powered code generation and analysis",
+  "description": "Shared utilities and types for AI-powered code generation, scaffolding, and analysis",
   "version": "0.1.0",
   "license": "AGPL-3.0",
   "author": "AgiflowIO",
@@ -15,7 +15,12 @@
   },
   "keywords": [
     "typescript",
-    "library"
+    "library",
+    "scaffold",
+    "generator",
+    "template",
+    "boilerplate",
+    "code-generation"
   ],
   "main": "./dist/index.cjs",
   "module": "./dist/index.js",
