+++ packages/aicode-utils/src/services/ScaffoldProcessingService.ts
