+++ packages/aicode-utils/src/services/index.ts
@@ -15,5 +15,6 @@
  * - Re-exporting types (types should come from '../types')
  */
 
-export * from './ProjectFinderService.js';
-export * from './TemplatesManagerService.js';
+export * from './ProjectFinderService';
+export * from './ScaffoldProcessingService';
+export * from './TemplatesManagerService';
