+++ packages/aicode-utils/tests/services/ScaffoldProcessingService.test.ts
