--- packages/scaffold-generator/CHANGELOG.md
@@ -1,48 +1,0 @@
-## 0.3.3 (2025-10-07)
-
-This was a version bump only for @agiflowai/scaffold-generator to align it with other projects, there were no code changes.
-
-## 0.3.1 (2025-10-06)
-
-This was a version bump only for @agiflowai/scaffold-generator to align it with other projects, there were no code changes.
-
-## 0.3.0 (2025-10-06)
-
-### 🚀 Features
-
-- **scaffold-mcp:** cli Update `add` command to get template from git sub folder. Update `init` command to pull all default templates. ([7b2da59](https://github.com/AgiFlow/aicode-toolkit/commit/7b2da59))
-- nextjs-15-drizzle template Add completed nextjs-15 template with drizzle and better-auth integration. ([8e21344](https://github.com/AgiFlow/aicode-toolkit/commit/8e21344))
-
-### ❤️ Thank You
-
-- Vuong Ngo @AgiFlow
-
-## 0.2.0 (2025-10-06)
-
-### 🚀 Features
-
-- **scaffold-mcp:** extra prompts Add two extra prompts for slash commands: scaffold-application and scaffold-feature ([54d1991](https://github.com/AgiFlow/aicode-toolkit/commit/54d1991))
-
-### ❤️ Thank You
-
-- Vuong Ngo @AgiFlow
-
-## 0.1.0 (2025-10-05)
-
-### 🚀 Features
-
-- nextjs-15 Add nextjs-15 template example ([f7512c5](https://github.com/AgiFlow/aicode-toolkit/commit/f7512c5))
-
-### ❤️ Thank You
-
-- Vuong Ngo @AgiFlow
-
-# 0.0.0 (2025-10-05)
-
-### 🚀 Features
-
-- init Initialize project ([adb2544](https://github.com/AgiFlow/aicode-toolkit/commit/adb2544))
-
-### ❤️ Thank You
-
-- Vuong Ngo
