--- packages/scaffold-generator/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@agiflowai/scaffold-generator",
-  "description": "Shared utilities and types for scaffold generators",
-  "version": "0.3.3",
-  "license": "AGPL-3.0",
-  "author": "AgiflowIO",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/AgiFlow/aicode-toolkit.git",
-    "directory": "packages/scaffold-generator"
-  },
-  "homepage": "https://github.com/AgiFlow/aicode-toolkit#readme",
-  "bugs": {
-    "url": "https://github.com/AgiFlow/aicode-toolkit/issues"
-  },
-  "keywords": [
-    "scaffold",
-    "generator",
-    "template",
-    "boilerplate",
-    "code-generation"
-  ],
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsdown",
-    "test": "vitest --run",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "fs-extra": "11.3.2"
-  },
-  "devDependencies": {
-    "@types/fs-extra": "^11.0.4",
-    "@types/node": "^22.0.0",
-    "@vitest/coverage-v8": "^3.0.0",
-    "tsdown": "^0.15.6",
-    "typescript": "5.9.3",
-    "vitest": "^3.0.0"
-  },
-  "type": "module",
-  "publishConfig": {
-    "access": "public"
-  }
-}
