--- packages/scaffold-generator/project.json
@@ -1,36 +1,0 @@
-{
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "name": "@agiflowai/scaffold-generator",
-  "projectType": "library",
-  "sourceRoot": "packages/scaffold-generator/src",
-  "targets": {
-    "build": {
-      "executor": "nx:run-commands",
-      "outputs": ["{projectRoot}/dist"],
-      "options": {
-        "command": "tsdown",
-        "cwd": "packages/scaffold-generator"
-      }
-    },
-    "typecheck": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "tsc --noEmit",
-        "cwd": "packages/scaffold-generator"
-      }
-    },
-    "lint": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "biome check packages/scaffold-generator"
-      }
-    },
-    "lint:fix": {
-      "executor": "nx:run-commands",
-      "options": {
-        "command": "biome check --write packages/scaffold-generator"
-      }
-    }
-  },
-  "tags": ["type:library", "scope:tools"]
-}
