--- packages/scaffold-generator/src/index.ts
@@ -1,2 +1,0 @@
-export * from './services';
-export * from './types';
