--- packages/scaffold-generator/tsdown.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from 'tsdown';
-
-export default defineConfig({
-  entry: ['src/index.ts'],
-  format: ['esm', 'cjs'],
-  clean: true,
-  shims: true,
-  dts: true,
-  packageExports: true,
-});
