+++ packages/scaffold-mcp/package.json
@@ -40,7 +40,6 @@
   },
   "dependencies": {
     "@agiflowai/aicode-utils": "workspace:*",
-    "@agiflowai/scaffold-generator": "workspace:*",
     "@composio/json-schema-to-zod": "0.1.15",
     "@modelcontextprotocol/sdk": "1.19.1",
     "chalk": "5.6.2",
